[+] More UI Elements [+] More convenience Functions [*] Fixed new challenges
New challenges does not work fully. It does reallocate the array and shift all references to the newly allocated array, and it will work without any issues if you reallocate the array to the same capacity. However, allocating it with larger capacity will throw an error due to set-sized arrays in other structures (ex ChallengeScreen, which has a set 70 buttons for each challenge and that cannot be changed.)